[Refactor] User, Jwt, OAuth2, Refrigerator 코드 개선
- /api/v1 삭제
- TokenUtil 내부메소드를 JwtPlugin 에 포함시켰습니다.
- User, Jwt, OAuth2, Refrigerator 내부 메서드에 대한 주석 설명을 추가했습니다.
- @PreAuthorize 추가
