+++ src/main/kotlin/team/b2/bingojango/domain/refrigerator/service/RefrigeratorService.kt
@@ -29,7 +29,7 @@ class RefrigeratorService(
         private val chatRoomService: ChatRoomService,
         private val mailRepository: MailRepository,
 ) {
-    //냉장고 목록 조회
+    //[API] 냉장고 목록 조회
     fun getRefrigerator(userPrincipal: UserPrincipal): List<RefrigeratorResponse> {
         val member = memberRepository.findAllByUserId(userPrincipal.id)
         val refrigerator = member.map { it.refrigerator }
@@ -37,7 +37,13 @@ class RefrigeratorService(
         return filtered.map { it.toResponse() }
     }
 
-    //신규 냉장고 생성
+    //[API] 신규 냉장고 생성
+    //1. 냉장고 이름 중복 여부 확인
+    //2. 비밀번호와 비밀번호확인의 일치 여부
+    //3. 냉장고 생성
+    //4. 채팅방 생성
+    //5. 멤버 생성 후, 생성한 사람에게 STAFF 권한 부여
+    //6. RefrigeratorResponse 반환
     @Transactional
     fun addRefrigerator(userPrincipal: UserPrincipal, request: AddRefrigeratorRequest): RefrigeratorResponse {
         if (refrigeratorRepository.existsRefrigeratorByName(request.name)) throw DuplicateValueException("중복된 냉장고 이름 입니다.")
@@ -49,14 +55,18 @@ class RefrigeratorService(
         return refrigerator.toResponse()
     }
 
-    //기존 냉장고 참여 - 비밀번호 이용
+    //[API] 기존 냉장고 참여 - 비밀번호 이용
+    //1. 냉장고 존재 유무 확인
+    //2. 냉장고 비밀번호 일치 여부 확인
+    //3. 채팅방 참여
+    //4. Member 권한으로 멤버로 참여
+    //5. RefrigeratorResponse 반환
     @Transactional
     fun joinRefrigeratorByPassword(userPrincipal: UserPrincipal, request: JoinByPasswordRequest): RefrigeratorResponse {
         val user = userRepository.findByIdOrNull(userPrincipal.id) ?: throw ModelNotFoundException("User")
-        //확인사항1: 냉장고 존재 유무
-        val refrigerator = refrigeratorRepository.findByName(request.name)
-                ?: throw ModelNotFoundException("Refrigerator")
-        //확인사항2: 비밀번호 일치 여부
+        //확인사항1:
+        val refrigerator = refrigeratorRepository.findByName(request.name) ?: throw ModelNotFoundException("Refrigerator")
+        //확인사항2:
         if (refrigerator.password != request.password) throw IllegalArgumentException("냉장고의 비밀번호가 일치하지 않습니다.")
 
         val chatRoom = chatRoomService.getChatRoom(refrigerator)
@@ -65,7 +75,13 @@ class RefrigeratorService(
         return refrigerator.toResponse()
     }
 
-    //기존 냉장고 참여 - 초대코드 이용
+    //[API] 기존 냉장고 참여 - 초대코드 이용
+    //1. 유저 확인
+    //2. 메일 코드 확인
+    //3. 냉장고 참여
+    //4. 채팅방 참여
+    //5. Member 권한으로 멤버로 참여
+    //6. RefrigeratorResponse 반환
     @Transactional
     fun joinRefrigeratorByInvitationCode(userPrincipal: UserPrincipal, request: JoinByInvitationCodeRequest): RefrigeratorResponse {
         val user = userRepository.findByIdOrNull(userPrincipal.id) ?: throw ModelNotFoundException("User")
