+++ src/main/kotlin/team/b2/bingojango/global/oauth/api/oauth2login/controller/OAuth2LoginController.kt
@@ -3,6 +3,7 @@ package team.b2.bingojango.global.oauth.api.oauth2login.controller
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import jakarta.servlet.http.HttpServletResponse
+import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RequestParam
@@ -18,16 +19,17 @@ class OAuth2LoginController(
     private val oAuth2LoginService: OAuth2LoginService,
     private val oAuth2ClientService: OAuth2ClientService
 ) {
-    //유저가 소셜로그인 요청 시, 로그인 페이지로 Redirect 해준다.
-    @Operation(summary = "로그인 페이지로 Redirect 해주기")
+        @Operation(summary = "로그인 페이지로 Redirect 해주기")
+    @PreAuthorize("isAnonymous()")
     @GetMapping("/oauth2/login/{provider}")
     fun redirectLoginPage(@PathVariable provider: OAuth2Provider, response: HttpServletResponse) {
         val loginPageUrl = oAuth2ClientService.generateLoginPageUrl(provider)
         response.sendRedirect(loginPageUrl)
     }
 
-    //카카오에게 받은 인증코드로 회원정보를 얻고, 유저에게 우리 서버의 access token을 반환한다.
-    @Operation(summary = "authorizationCode ")
+
+    @Operation(summary = "provider 인증 후 AccessToken 반환")
+    @PreAuthorize("isAnonymous()")
     @GetMapping("/oauth2/callback/{provider}")
     fun callback(
         @PathVariable provider: OAuth2Provider, response: HttpServletResponse,
