+++ src/main/kotlin/team/b2/bingojango/global/security/jwt/JwtAuthenticationToken.kt
@@ -5,6 +5,7 @@ import org.springframework.security.web.authentication.WebAuthenticationDetails
 import team.b2.bingojango.global.security.util.UserPrincipal
 import java.io.Serializable
 
+//사용자의 인증 정보와 권한 처리
 class JwtAuthenticationToken(
     private val principal: UserPrincipal,
     details: WebAuthenticationDetails?,
@@ -14,6 +15,7 @@ class JwtAuthenticationToken(
         super.setAuthenticated(true)
         if (details != null)super.setDetails(details)
     }
+
     override fun getPrincipal() = principal
     override fun getCredentials() = null
     override fun isAuthenticated(): Boolean {
