+++ src/main/kotlin/team/b2/bingojango/global/security/jwt/JwtPlugin.kt
@@ -6,10 +6,14 @@ import io.jsonwebtoken.Jwts
 import io.jsonwebtoken.security.Keys
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.PropertySource
+import org.springframework.data.repository.findByIdOrNull
 import org.springframework.security.oauth2.core.user.OAuth2User
 import org.springframework.stereotype.Component
+import org.springframework.transaction.annotation.Transactional
+import team.b2.bingojango.domain.user.model.User
 import team.b2.bingojango.domain.user.repository.UserRepository
 import team.b2.bingojango.global.exception.cases.ModelNotFoundException
+import team.b2.bingojango.global.security.util.UserPrincipal
 import java.nio.charset.StandardCharsets
 import java.time.Duration
 import java.time.Instant
@@ -23,31 +27,25 @@ class JwtPlugin(
     @Value("\${auth.jwt.accessTokenExpirationHour}") private val accessTokenExpirationHour: Long,
     @Value("\${auth.jwt.refreshTokenExpirationHour}") private val refreshTokenExpirationHour: Long,
     private val userRepository: UserRepository,
+    private val tokenRepository: TokenRepository,
 ) {
-    fun validateToken(jwt: String): Result<Jws<Claims>>{
-        return kotlin.runCatching {
-            val key = Keys.hmacShaKeyFor(secret.toByteArray(StandardCharsets.UTF_8))
-            Jwts.parser().verifyWith(key).build().parseSignedClaims(jwt)
-        }
-    }
-
+    //AccessToken 생성
     fun generateAccessToken(subject: String, email: String, role: String): String {
         return generateToken(subject, email, role, Duration.ofHours(accessTokenExpirationHour))
     }
 
+    //RefreshToken 생성
     fun generateRefreshToken(subject: String, email: String, role: String): String {
         return generateToken(subject, email, role, Duration.ofHours(refreshTokenExpirationHour))
     }
 
+    //Token 생성
     private fun generateToken(subject: String, email: String, role: String, expirationPeriod: Duration): String {
         val claims: Claims = Jwts.claims()
             .add(mapOf("role" to role, "email" to email))
             .build()
-
         val key = Keys.hmacShaKeyFor(secret.toByteArray(StandardCharsets.UTF_8))
-
         val now = Instant.now()
-
         return Jwts.builder()
             .subject(subject)
             .issuer(issuer)
@@ -58,6 +56,7 @@ class JwtPlugin(
             .compact()
     }
 
+    //Token Dto 생성
     fun generateTokenDto(oAuth2User: OAuth2User): JwtDto {
         val email = oAuth2User.attributes["email"] as String
         val user = userRepository.findByEmail(email) ?: throw ModelNotFoundException("User")
@@ -65,11 +64,31 @@ class JwtPlugin(
         val role = user.role.toString()
         val refreshToken = generateRefreshToken(subject, email, role)
         val accessToken = generateAccessToken(subject, email, role)
-
         return JwtDto(
             accessToken = accessToken,
             refreshToken = refreshToken,
         )
     }
 
+    //토큰 검증
+    fun validateToken(jwt: String): Result<Jws<Claims>>{
+        return kotlin.runCatching {
+            val key = Keys.hmacShaKeyFor(secret.toByteArray(StandardCharsets.UTF_8))
+            Jwts.parser().verifyWith(key).build().parseSignedClaims(jwt)
+        }
+    }
+
+    //토큰 저장
+    @Transactional
+    fun storeToken(user: User, refreshToken: String) {
+        tokenRepository.save(RefreshToken(user, refreshToken))
+    }
+
+    //토큰 삭제
+    @Transactional
+    fun deleteToken(userPrincipal: UserPrincipal) {
+        val user = userRepository.findByIdOrNull(userPrincipal.id) ?: throw ModelNotFoundException ("User")
+        val refreshToken = tokenRepository.findByUser(user) ?: throw ModelNotFoundException("RefreshToken")
+        tokenRepository.delete(refreshToken)
+    }
 }
